+++ backend/src/domains/products/public.types.ts
@@ -2,14 +2,14 @@ import { ProductEntity } from './main/entities/product.entity';
 
 export type TProducts = ProductEntity;
 
-export type TCreateProducts = Omit<
+export type TCreateProduct = Omit<
   ProductEntity,
   'id' | 'createdAt' | 'updatedAt'
 > & {
   image: Express.Multer.File;
 };
 
-export type TUpdateProducts = Partial<
+export type TUpdateProduct = Partial<
   Omit<ProductEntity, 'id' | 'createdAt' | 'updatedAt'>
 > & {
   image?: Express.Multer.File;