+++ backend/src/gateways/admin/sellers/sellers.controller.ts
@@ -3,6 +3,10 @@ import {
   Controller,
   Get,
   Post,
+  Delete,
+  Patch,
+  Param,
+  ParseIntPipe,
   UploadedFile,
   UseInterceptors,
 } from '@nestjs/common';
@@ -10,6 +14,7 @@ import { AD } from '../admin.const';
 import { SellersPublicService } from 'src/domains/sellers/public.service';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { CreateSellerDto } from './dto/createSellerDto';
+import { UpdateSellerDto } from './dto/updateSellerDto';
 
 @Controller(`${AD}/sellers`)
 export class SellersController {
@@ -27,5 +32,20 @@ export class SellersController {
   @Get('')
   async findAll() {
     return this.sellersPublicService.findAll();
+  }
+
+  @Patch(':id')
+  @UseInterceptors(FileInterceptor('image'))
+  async update(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() body: UpdateSellerDto,
+    @UploadedFile() image?: Express.Multer.File,
+  ) {
+    return this.sellersPublicService.update(id, { ...body, image });
+  }
+
+  @Delete(':id')
+  async delete(@Param('id', ParseIntPipe) id: number) {
+    return this.sellersPublicService.delete(id);
   }
 }